refactor(services): split mercury/protocol.rs into submodules
Promote protocol.rs (754 lines) into a directory split by lifecycle:
session (connect/sync/teardown), character (char-list/visuals), and
resources (cooked-data fragments and version info). Tests live in tests.rs.

Public API unchanged. No behavior changes.
